lablink-allocator/requirements.txt
Flask==2.2.2
werkzeug==2.2.2
flask_sqlalchemy==3.0.0
psycopg2-binary==2.9.6

